+++ apollo-configservice/src/main/java/com/ctrip/framework/apollo/metaservice/controller/HomePageController.java
@@ -29,7 +29,14 @@ import org.springframework.web.bind.annotation.RestController;
 /**
  * For non-eureka discovery services such as kubernetes and nacos, there is no eureka home page, so we need to add a default one
  */
-@Profile({"kubernetes", "nacos-discovery", "consul-discovery", "zookeeper-discovery"})
+@Profile({
+    "kubernetes",
+    "nacos-discovery",
+    "consul-discovery",
+    "zookeeper-discovery",
+    "custom-defined-discovery",
+    "database-discovery",
+})
 @RestController
 public class HomePageController {
   private final DiscoveryService discoveryService;