+++ apollo-configservice/src/main/java/com/ctrip/framework/apollo/metaservice/service/DefaultDiscoveryService.java
@@ -33,7 +33,14 @@ import org.springframework.util.CollectionUtils;
  * Default discovery service for Eureka
  */
 @Service
-@ConditionalOnMissingProfile({"kubernetes", "nacos-discovery", "consul-discovery", "zookeeper-discovery", "custom-defined-discovery"})
+@ConditionalOnMissingProfile({
+    "kubernetes",
+    "nacos-discovery",
+    "consul-discovery",
+    "zookeeper-discovery",
+    "custom-defined-discovery",
+    "database-discovery",
+})
 public class DefaultDiscoveryService implements DiscoveryService {
 
   private final EurekaClient eurekaClient;