+++ CHANGES.md
@@ -33,6 +33,7 @@ Apollo 2.1.0
 * [Move apollo-core, apollo-client, apollo-mockserver, apollo-openapi and apollo-client-config-data to apollo-java repo](https://github.com/apolloconfig/apollo/pull/4594)
 * [fix get the openapi interface that contains namespace information for deleted items](https://github.com/apolloconfig/apollo/pull/4596)
 * [A user-friendly config management page for apollo portal](https://github.com/apolloconfig/apollo/pull/4592)
+* [feat: support use database as a registry](https://github.com/apolloconfig/apollo/pull/4595)
 
 ------------------
 All issues and pull requests are [here](https://github.com/apolloconfig/apollo/milestone/11?closed=1)